+++ setup.py
@@ -83,9 +83,9 @@ with codecs.open('README.rst', 'r', 'utf-8') as fd:
       package_data={
           'yapf': [
               'third_party/yapf_diff/LICENSE',
-              'third_party/ylib2to3/Grammar.txt',
-              'third_party/ylib2to3/PatternGrammar.txt',
-              'third_party/ylib2to3/LICENSE',
+              'third_party/_ylib2to3/Grammar.txt',
+              'third_party/_ylib2to3/PatternGrammar.txt',
+              'third_party/_ylib2to3/LICENSE',
           ]
       },
       include_package_data=True,
