+++ yapf/pytree/pytree_utils.py
@@ -27,8 +27,8 @@ the lib2to3 library.
 import ast
 import os
 
-from yapf.third_party.ylib2to3 import pygram, pytree
-from yapf.third_party.ylib2to3.pgen2 import driver, parse, token
+from yapf.third_party._ylib2to3 import pygram, pytree
+from yapf.third_party._ylib2to3.pgen2 import driver, parse, token
 
 # TODO(eliben): We may want to get rid of this filtering at some point once we
 # have a better understanding of what information we need from the tree. Then,
