+++ yapf/pytree/pytree_visitor.py
@@ -26,7 +26,7 @@ a pytree into a stream.
 
 import sys
 
-from yapf.third_party.ylib2to3 import pytree
+from yapf.third_party._ylib2to3 import pytree
 
 from yapf.pytree import pytree_utils
