+++ yapf/pytree/split_penalty.py
@@ -15,8 +15,8 @@
 
 import re
 
-from yapf.third_party.ylib2to3 import pytree
-from yapf.third_party.ylib2to3.pgen2 import token as grammar_token
+from yapf.third_party._ylib2to3 import pytree
+from yapf.third_party._ylib2to3.pgen2 import token as grammar_token
 
 from yapf.pytree import pytree_utils
 from yapf.pytree import pytree_visitor
