+++ yapf/yapflib/errors.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 """YAPF error objects."""
 
-from yapf.third_party.ylib2to3.pgen2 import tokenize
+from yapf.third_party._ylib2to3.pgen2 import tokenize
 
 
 def FormatErrorMsg(e):
