+++ yapf/yapflib/logical_line.py
@@ -19,7 +19,7 @@ line if there were no line length restrictions. It's then used by the parser to
 perform the wrapping required to comply with the style guide.
 """
 
-from yapf.third_party.ylib2to3.fixer_util import syms as python_symbols
+from yapf.third_party._ylib2to3.fixer_util import syms as python_symbols
 
 from yapf.pytree import pytree_utils
 from yapf.pytree import split_penalty
