+++ yapf/yapflib/reformatter.py
@@ -23,8 +23,8 @@ import collections
 import heapq
 import re
 
-from yapf.third_party.ylib2to3 import pytree
-from yapf.third_party.ylib2to3.pgen2 import token
+from yapf.third_party._ylib2to3 import pytree
+from yapf.third_party._ylib2to3.pgen2 import token
 
 from yapf.pytree import pytree_utils
 from yapf.yapflib import format_decision_state
