+++ yapftests/logical_line_test.py
@@ -16,8 +16,8 @@
 import textwrap
 import unittest
 
-from yapf.third_party.ylib2to3 import pytree
-from yapf.third_party.ylib2to3.pgen2 import token
+from yapf.third_party._ylib2to3 import pytree
+from yapf.third_party._ylib2to3.pgen2 import token
 
 from yapf.pytree import split_penalty
 from yapf.yapflib import format_token
