+++ yapftests/pytree_utils_test.py
@@ -15,9 +15,9 @@
 
 import unittest
 
-from yapf.third_party.ylib2to3 import pygram
-from yapf.third_party.ylib2to3 import pytree
-from yapf.third_party.ylib2to3.pgen2 import token
+from yapf.third_party._ylib2to3 import pygram
+from yapf.third_party._ylib2to3 import pytree
+from yapf.third_party._ylib2to3.pgen2 import token
 
 from yapf.pytree import pytree_utils
