+++ yapftests/yapf_test.py
@@ -25,7 +25,7 @@ import textwrap
 import unittest
 
 from io import StringIO
-from yapf.third_party.ylib2to3.pgen2 import tokenize
+from yapf.third_party._ylib2to3.pgen2 import tokenize
 
 from yapf.yapflib import errors
 from yapf.yapflib import style
